pyproject.toml
[project]
name = "student-guider-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "MuhammadAbdullah95", email = "ma2404374@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.13",
    "docx2txt>=0.9",
    "fastapi[standard]>=0.115.14",
    "google-genai>=1.24.0",
    "langchain-community>=0.3.26",
    "openai-agents>=0.1.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
]

[project.scripts]
student-guider-chatbot = "student_guider_chatbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
